Extract an AbstractSubmitTxCommand
