+++ exonum-java-binding/tutorials/car-registry/car-registry-client/src/main/java/com/example/car/client/AddVehicleCommand.java
@@ -18,11 +18,9 @@ package com.example.car.client;
 
 import com.example.car.messages.Transactions;
 import com.example.car.messages.VehicleOuterClass.Vehicle;
-import com.exonum.binding.common.crypto.CryptoFunctions;
+import com.exonum.binding.common.crypto.KeyPair;
 import com.exonum.binding.common.message.TransactionMessage;
-import com.exonum.client.ExonumClient;
 import java.util.concurrent.Callable;
-import picocli.CommandLine.ArgGroup;
 import picocli.CommandLine.Command;
 import picocli.CommandLine.Parameters;
 
@@ -32,9 +30,6 @@ import picocli.CommandLine.Parameters;
 public final class AddVehicleCommand implements Callable<Integer> {
 
   private static final int ADD_VEHICLE_TX_ID = 0;
-
-  @ArgGroup(exclusive = true, multiplicity = "1")
-  ServiceIds serviceIds;
 
   @Parameters(index = "0")
   String id;
@@ -49,17 +44,9 @@ public final class AddVehicleCommand implements Callable<Integer> {
   String owner;
 
   @Override
-  public Integer call() {
-    var client = ExonumClient.newBuilder()
-        .setExonumHost(Config.NODE_PUBLIC_API_HOST)
-        .build();
-
-    // todo (here and in other txs): Shall we add a command to generate a keypair and let all
-    //   other use that keypair, so that the user learns at least that each transaction comes
-    //   signed with a key into the network?
-    var keyPair = CryptoFunctions.ed25519().generateKeyPair();
-    var txMessage = TransactionMessage.builder()
-        .serviceId(findServiceId(client))
+  protected TransactionMessage createTxMessage(int serviceId, KeyPair keyPair) {
+    return TransactionMessage.builder()
+        .serviceId(serviceId)
         .transactionId(ADD_VEHICLE_TX_ID)
         .payload(
             Transactions.AddVehicle.newBuilder()
@@ -71,14 +58,5 @@ public final class AddVehicleCommand implements Callable<Integer> {
                         .setOwner(owner))
                 .build())
         .sign(keyPair);
-
-    // todo: Add logging
-    var txHashCode = client.submitTransaction(txMessage);
-    return 0;
-  }
-
-  private int findServiceId(ExonumClient client) {
-    var serviceIdResolver = new ServiceIdResolver(serviceIds, client);
-    return serviceIdResolver.getId();
   }
 }
